Hack up an initial X version.
This is a copy of the console version, modified to use the XRECORD
extension. It doesn't currently scan the keyboard on startup, and
the numlock handling isn't there. X doesn't appear to have autorepeat
functionality either.  Finally, the xkeymap.txt generation is pretty
cursory... in particular, the Super keys are not handled.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>